+++ src/helpers/config.ts
@@ -2,4 +2,5 @@ export const POSSIBLE_REPORT_EXTENSIONS = [".html"];
 export const POSSIBLE_TRACE_EXTENSIONS = [".zip"];
 export const POSSIBLE_SPEC_EXTENSIONS = [".spec.ts", ".test.ts"];
 export const DEFAULT_TRACE_FILE_NAME = "trace.zip";
+export const DEFAULT_TRACE_FILE_PARTS = ["trace", "zip"];
 export const DEFAULT_REPORT_FILE_NAME = "index.html";